requirements.txt
python-telegram-bot>=21.0.0
Pillow>=10.0.0
requests>=2.31.0
Flask>=2.3.0
Flask-CORS>=4.0.0

